Изменены примитивные типы на ссылочные, доработаны методы работы с БД

+++ src/main/java/entity/ProductsWithSales.java
@@ -4,9 +4,9 @@ import lombok.Data;
 
 @Data
 public class ProductsWithSales {
-    private int productId;
+    private Long productId;
     private String productName;
     private String category;
-    private int quantity;
-    private double price;
+    private Integer quantity;
+    private Double price;
 }
